backend/requirements.txt
fastapi
uvicorn
python-multipart
pydub
requests
python-dotenv
vosk
